// packages/protocol/typescript/generated/yiru/runtime/v1/browser_pb.d.ts
export declare type BrowserValue = Message<"yiru.runtime.v1.BrowserValue"> & {
  /**
   * @generated from oneof yiru.runtime.v1.BrowserValue.kind
   */
  kind: {
    /**
     * @generated from field: yiru.runtime.v1.BrowserNull null_value = 1;
     */
    value: BrowserNull;
    case: "nullValue";
  } | {
    /**
     * @generated from field: bool bool_value = 2;
     */
    value: boolean;
    case: "boolValue";
  } | {
    /**
     * @generated from field: double number_value = 3;
     */
    value: number;
    case: "numberValue";
  } | {
    /**
     * @generated from field: string string_value = 4;
     */
    value: string;
    case: "stringValue";
  } | {
    /**
     * @generated from field: yiru.runtime.v1.BrowserValueList list_value = 5;
     */
    value: BrowserValueList;
    case: "listValue";
  } | {
    /**
     * @generated from field: yiru.runtime.v1.BrowserValueObject object_value = 6;
     */
    value: BrowserValueObject;
    case: "objectValue";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message yiru.runtime.v1.BrowserValue.
 * Use `create(BrowserValueSchema)` to create a new message.
 */
export declare const BrowserValueSchema: GenMessage<BrowserValue>;

/**
 * @generated from message yiru.runtime.v1.BrowserValueList
 */
export declare type BrowserValueList = Message<"yiru.runtime.v1.BrowserValueList"> & {
  /**
   * @generated from field: repeated yiru.runtime.v1.BrowserValue values = 1;
   */
  values: BrowserValue[];
};

/**
 * Describes the message yiru.runtime.v1.BrowserValueList.
 * Use `create(BrowserValueListSchema)` to create a new message.
 */
export declare const BrowserValueListSchema: GenMessage<BrowserValueList>;

/**
 * @generated from message yiru.runtime.v1.BrowserValueObject
 */
export declare type BrowserValueObject = Message<"yiru.runtime.v1.BrowserValueObject"> & {
  /**
   * @generated from field: repeated yiru.runtime.v1.BrowserValueEntry entries = 1;
   */
  entries: BrowserValueEntry[];
};

/**
 * Describes the message yiru.runtime.v1.BrowserValueObject.
 * Use `create(BrowserValueObjectSchema)` to create a new message.
 */
export declare const BrowserValueObjectSchema: GenMessage<BrowserValueObject>;

/**
 * @generated from message yiru.runtime.v1.BrowserValueEntry
 */
export declare type BrowserValueEntry = Message<"yiru.runtime.v1.BrowserValueEntry"> & {
  /**
   * @generated from field: string key = 1;
   */
  key: string;

  /**
   * @generated from field: yiru.runtime.v1.BrowserValue value = 2;
   */
  value?: BrowserValue | undefined;
};

/**
 * Describes the message yiru.runtime.v1.BrowserValueEntry.
 * Use `create(BrowserValueEntrySchema)` to create a new message.
 */
export declare const BrowserValueEntrySchema: GenMessage<BrowserValueEntry>;

/**
 * @generated from enum yiru.runtime.v1.BrowserNull
 */
export enum BrowserNull {
  /**
   * @generated from enum value: BROWSER_NULL_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: BROWSER_NULL_VALUE = 1;
   */
  VALUE = 1,
}